Make a package for navigation routes
